# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/ai/{providerId}/models:
        get:
            tags:
                - AIService
            description: "ListProviderModels lists the models an instance AI provider offers. The\r\n server calls the provider's model catalog endpoint, so this requires\r\n network access to the provider."
            operationId: AIService_ListProviderModels
            parameters:
                - name: providerId
                  in: path
                  description: Required. The provider id from the instance AI setting.
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListProviderModelsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:chat:
        post:
            tags:
                - AIService
            description: "Chat runs one conversational turn with an instance AI provider. The caller\r\n resends the whole conversation each turn; the server keeps no chat state.\r\n The model may propose note changes, but this method never writes a memo."
            operationId: AIService_Chat
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ChatRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ChatResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:estimateChatContext:
        post:
            tags:
                - AIService
            description: "EstimateChatContext resolves a context selection and reports how much of\r\n the chat context budget it would consume. The Hub calls this while the user\r\n edits the selection so an over-budget selection is refused before sending."
            operationId: AIService_EstimateChatContext
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EstimateChatContextRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EstimateChatContextResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:transcribe:
        post:
            tags:
                - AIService
            description: Transcribe transcribes an audio file using an instance AI provider.
            operationId: AIService_Transcribe
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TranscribeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TranscribeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
        AIProviderModel:
            type: object
            properties:
                id:
                    type: string
                    description: The provider-specific model identifier, e.g. "anthropic/claude-sonnet-4".
                contextLength:
                    type: string
                    description: "The model's total context window in tokens. Zero when the provider does not\r\n report one."
...
        ChatMessage:
            type: object
            properties:
                role:
                    enum:
                        - CHAT_MESSAGE_ROLE_UNSPECIFIED
                        - CHAT_MESSAGE_ROLE_USER
                        - CHAT_MESSAGE_ROLE_ASSISTANT
                    type: string
                    format: enum
                content:
                    type: string
                    description: "The message text. For assistant messages this is the visible reply with any\r\n proposal blocks removed."
        ChatProposal:
            type: object
            properties:
                action:
                    enum:
                        - CHAT_PROPOSAL_ACTION_UNSPECIFIED
                        - CHAT_PROPOSAL_ACTION_CREATE
                        - CHAT_PROPOSAL_ACTION_UPDATE
                    type: string
                    format: enum
                content:
                    type: string
                    description: "The proposed note content in Markdown. Tags are derived from #hashtags in\r\n this content, matching CreateMemo."
                target:
                    type: string
                    description: The note to update, formatted "memos/{uid}". Empty for create proposals.
                targetContent:
                    type: string
                    description: |-
                        The target note content the model read. Present only when an update target
                         belongs to this turn's selected context; clients use it as an optimistic
                         concurrency precondition when applying the proposal.
            description: "ChatProposal is a note change the model suggested but did not perform. The\r\n Hub shows it for confirmation and applies it through the normal memo API, so\r\n access control and change notifications stay in one place."
        ChatRequest:
            required:
                - messages
            type: object
            properties:
                filter:
                    type: string
                    description: "Optional. A CEL filter selecting which notes the model may read. An empty\r\n filter means the model sees none of the user's notes."
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/ChatMessage'
                    description: "Required. The conversation so far, oldest first, ending with the new user\r\n message. The server keeps no conversation state."
        ChatResponse:
            type: object
            properties:
                content:
                    type: string
                    description: The model's reply, with proposal blocks removed.
                proposals:
                    type: array
                    items:
                        $ref: '#/components/schemas/ChatProposal'
                    description: The note changes the model proposed. Never applied by the server.
                contextMemoCount:
                    type: string
                    description: How many notes were injected as context.
                contextEstimatedTokens:
                    type: string
                    description: The estimated token cost of the injected context.
                contextBudgetTokens:
                    type: string
                    description: The context budget this turn was measured against.
                promptTokens:
                    type: string
                    description: Token usage reported by the provider. Zero when the provider omits it.
                completionTokens:
                    type: string
                totalTokens:
                    type: string
                model:
                    type: string
                    description: "The model the provider reports having used, which can differ from the\r\n requested model when a router picks a backend."
                truncated:
                    type: boolean
                    description: True when the provider stopped because the reply hit the token limit.
...
        EstimateChatContextRequest:
            required:
                - filter
            type: object
            properties:
                filter:
                    type: string
                    description: "Required. A CEL filter selecting which notes the model may read. An empty\r\n filter selects nothing, not everything."
        EstimateChatContextResponse:
            type: object
            properties:
                memoCount:
                    type: string
                    description: How many notes the selection matched.
                totalChars:
                    type: string
                    description: The total character count of the matched notes' content.
                estimatedTokens:
                    type: string
                    description: The estimated token cost of the selection.
                contextBudgetTokens:
                    type: string
                    description: The budget this estimate was measured against.
                fits:
                    type: boolean
                    description: Whether the selection fits the budget.
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListProviderModelsResponse:
            type: object
            properties:
                models:
                    type: array
                    items:
                        $ref: '#/components/schemas/AIProviderModel'
                    description: The models the provider offers, ordered by identifier.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TranscribeRequest:
            required:
                - audio
            type: object
            properties:
                audio:
                    allOf:
                        - $ref: '#/components/schemas/TranscriptionAudio'
                    description: Required. Audio input.
        TranscribeResponse:
            type: object
            properties:
                text:
                    type: string
                    description: The transcribed text.
        TranscriptionAudio:
            type: object
            properties:
                content:
                    writeOnly: true
                    type: string
                    description: Inline audio bytes.
                    format: bytes
                uri:
                    type: string
                    description: URI for audio content. Reserved for future use.
                filename:
                    type: string
                    description: Optional. The uploaded filename.
                contentType:
                    type: string
                    description: Optional. The MIME type of the input audio.
...